+++ src/stdune/loc.mli
@@ -2,6 +2,8 @@ include module type of struct
   include Loc0
 end
 
+module Map : Map_intf.S with type key := t
+
 val in_file : Path.t -> t
 
 val in_dir : Path.t -> t
@@ -25,8 +27,16 @@ val of_pos : string * int * int * int -> t
 
 val to_file_colon_line : t -> string
 
-val pp_file_colon_line : Format.formatter -> t -> unit
+val pp_file_colon_line : t -> unit Pp.t
 
-val print : Format.formatter -> t -> unit
+val to_dyn_hum : t -> Dyn.t
+
+type tag = Loc
+
+val pp : t -> tag Pp.t
+
+val render : Format.formatter -> tag Pp.t -> unit
 
 val on_same_line : t -> t -> bool
+
+val span : t -> t -> t